% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/departamentos.R
\name{agregar_departamento}
\alias{agregar_departamento}
\title{Agregar columna de departamento a datos EPHC}
\usage{
agregar_departamento(data)
}
\arguments{
\item{data}{data.frame de ine_ephc(). Requiere variable ESTGEO.}
}
\value{
El mismo data.frame con una columna adicional
"departamento" con el nombre del departamento.
}
\description{
Deriva el nombre del departamento a partir de la variable ESTGEO
de la EPHC del INE de Paraguay. La variable ESTGEO codifica el
estrato geografico combinando departamento y area (urbano/rural).
Esta funcion es la base para todas las desagregaciones
departamentales. Nota: la EPHC trimestral no cubre los
departamentos de Boqueron y Alto Paraguay.
}
\examples{
\dontrun{
ephc <- ine_ephc(2024, 1)
ephc <- agregar_departamento(ephc)
table(ephc$departamento)
}
}
