% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abandono.R
\name{grafico_abandono_departamento}
\alias{grafico_abandono_departamento}
\title{Grafico de abandono por departamento}
\usage{
grafico_abandono_departamento(data, titulo = NULL)
}
\arguments{
\item{data}{data.frame cargado con mec_leer().}

\item{titulo}{Titulo del grafico. Opcional.}
}
\value{
Un objeto ggplot2.
}
\description{
Genera un grafico de barras horizontales mostrando la tasa de
abandono escolar por departamento. Permite identificar los
departamentos con mayor desercion escolar en Paraguay.
}
\examples{
\dontrun{
basica <- mec_leer("matriculaciones_educacion_escolar_basica.csv")
grafico_abandono_departamento(basica)
}
}
