% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ine.R
\name{ine_geo}
\alias{ine_geo}
\title{Descargar codigos geograficos del INE}
\usage{
ine_geo(nivel = "departamentos")
}
\arguments{
\item{nivel}{Nivel geografico a descargar. Opciones:
\describe{
\item{"departamentos"}{18 departamentos del Paraguay}
\item{"distritos"}{Distritos con codigo concatenado}
\item{"barrios"}{Barrios y localidades del Paraguay}
}}
}
\value{
Un data.frame con los codigos geograficos del nivel
seleccionado.
}
\description{
Descarga los codigos geograficos oficiales del Instituto Nacional
de Estadistica (INE) de Paraguay, utilizados en todos los
microdatos del INE para identificar departamentos, distritos
y barrios.
}
\examples{
\dontrun{
deptos <- ine_geo("departamentos")
distritos <- ine_geo("distritos")
barrios <- ine_geo("barrios")
ephc <- ine_ephc(2024, 1)
ephc_geo <- merge(ephc, deptos,
                  by.x = "ESTGEO",
                  by.y = "codigo_dpto")
}
}
