% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisa.R
\docType{data}
\name{pisa_paraguay}
\alias{pisa_paraguay}
\title{Resultados historicos PISA - Paraguay 2017 y 2022}
\format{
Un data.frame con 2 filas y 12 columnas:
\describe{
\item{anio}{Anio de aplicacion de la prueba}
\item{matematica}{Puntaje promedio en matematica}
\item{lectura}{Puntaje promedio en lectura}
\item{ciencias}{Puntaje promedio en ciencias}
\item{promedio_ocde_matematica}{Promedio OCDE en matematica}
\item{promedio_ocde_lectura}{Promedio OCDE en lectura}
\item{promedio_ocde_ciencias}{Promedio OCDE en ciencias}
\item{pct_bajo_rendimiento_matematica}{Porcentaje bajo nivel 2
en matematica}
\item{pct_bajo_rendimiento_lectura}{Porcentaje bajo nivel 2
en lectura}
\item{pct_bajo_rendimiento_ciencias}{Porcentaje bajo nivel 2
en ciencias}
\item{n_estudiantes}{Numero de estudiantes evaluados}
\item{n_escuelas}{Numero de escuelas participantes}
}
}
\source{
OCDE (2023). PISA 2022 Results (Volume I). OECD Publishing.
MEC Paraguay (2023). Informe Nacional PISA 2022.
}
\usage{
pisa_paraguay
}
\description{
Dataset con los puntajes promedio de Paraguay en las dos ediciones
de PISA en que ha participado (2017 y 2022). Incluye comparacion
con el promedio OCDE y porcentaje de estudiantes con bajo rendimiento
(por debajo del nivel 2 de competencia).
}
\keyword{datasets}
